Move `PendingPoint` and make it crate private
